refactor(#51): move convertBufferToTransferable
