+++ src/core/encryption/WorkerEncryptionController.ts
@@ -1,6 +1,7 @@
 import { WorkerMessenger } from '../../utils/workers/WorkerMessenger';
 import { WorkerRPC } from '../../utils/workers/WorkerRPC';
 
+import { convertBufferToTransferable } from './utils/buffers';
 import { ICipher } from '.';
 
 export class Terminable {
@@ -19,58 +20,6 @@ export class Terminable {
 		if (this.isFinished) throw new Error(errorMessage);
 	}
 }
-
-/**
- * Receive any buffer and returns a converted transferable buffer and codec for revert converting
- *
- * Useful to convert buffer for transfer and then convert received buffer back.
- *
- * @param buffer
- * @returns [buffer, convertor]
- */
-export const convertBufferToTransferable = (
-	buffer: ArrayBuffer,
-): [ArrayBuffer, (buffer: ArrayBuffer) => ArrayBuffer] => {
-	if (buffer instanceof Buffer) {
-		return [
-			buffer.buffer,
-			(buffer: ArrayBuffer) => {
-				return Buffer.from(buffer);
-			},
-		];
-	}
-
-	// TODO: add tests
-	// Typed arrays
-	const typedArray = [
-		Int8Array,
-		Uint8Array,
-		Uint8ClampedArray,
-		Int16Array,
-		Uint16Array,
-		Int32Array,
-		Uint32Array,
-		Float32Array,
-		Float64Array,
-		BigInt64Array,
-		BigUint64Array,
-	].find((proto) => buffer instanceof proto);
-	if (typedArray && buffer instanceof typedArray) {
-		return [
-			buffer.buffer,
-			(buffer: ArrayBuffer) => {
-				return new typedArray(buffer);
-			},
-		];
-	}
-
-	return [
-		buffer,
-		(buffer: ArrayBuffer) => {
-			return buffer;
-		},
-	];
-};
 
 export class WorkerEncryptionController implements ICipher {
 	private readonly worker;
