+++ src/core/encryption/utils/buffers.ts
@@ -57,3 +57,55 @@ export class BufferView {
 		}
 	}
 }
+
+/**
+ * Receive any buffer and returns a converted transferable buffer and codec for revert converting
+ *
+ * Useful to convert buffer for transfer and then convert received buffer back.
+ *
+ * @param buffer
+ * @returns [buffer, convertor]
+ */
+export const convertBufferToTransferable = (
+	buffer: ArrayBuffer,
+): [ArrayBuffer, (buffer: ArrayBuffer) => ArrayBuffer] => {
+	if (buffer instanceof Buffer) {
+		return [
+			buffer.buffer,
+			(buffer: ArrayBuffer) => {
+				return Buffer.from(buffer);
+			},
+		];
+	}
+
+	// TODO: add tests
+	// Typed arrays
+	const typedArray = [
+		Int8Array,
+		Uint8Array,
+		Uint8ClampedArray,
+		Int16Array,
+		Uint16Array,
+		Int32Array,
+		Uint32Array,
+		Float32Array,
+		Float64Array,
+		BigInt64Array,
+		BigUint64Array,
+	].find((proto) => buffer instanceof proto);
+	if (typedArray && buffer instanceof typedArray) {
+		return [
+			buffer.buffer,
+			(buffer: ArrayBuffer) => {
+				return new typedArray(buffer);
+			},
+		];
+	}
+
+	return [
+		buffer,
+		(buffer: ArrayBuffer) => {
+			return buffer;
+		},
+	];
+};
